requirements.txt
streamlit>=1.32.0
pandas>=2.1.4
numpy<2.0.0
plotly>=5.18.0
scikit-learn>=1.3.2
xgboost>=2.0.3
tensorflow-cpu==2.15.0
yfinance>=0.2.64
ta>=0.10.2
joblib>=1.3.2
requests
